+++ packages/cad-viewer-example/vite.config.ts
@@ -44,7 +44,7 @@ export default defineConfig(({ command, mode }) => {
           dest: 'assets'
         },
         {
-          src: './node_modules/@mlightcad/cad-simple-viewer/dist/libredwg-parser-worker.js',
+          src: './node_modules/@mlightcad/cad-simple-viewer/dist/*-worker.js',
           dest: 'assets'
         }
       ]