+++ packages/three-renderer/src/object/AcTrMText.ts
@@ -1,12 +1,15 @@
 import { AcGiMTextData, AcGiTextStyle } from '@mlightcad/data-model'
-import { FontManager, MText } from '@mlightcad/mtext-renderer'
+import { MTextObject } from '@mlightcad/mtext-renderer'
 import * as THREE from 'three'
 
+import { AcTrMTextRenderer } from '../renderer'
 import { AcTrStyleManager } from '../style/AcTrStyleManager'
 import { AcTrEntity } from './AcTrEntity'
 
 export class AcTrMText extends AcTrEntity {
-  private _mtext?: MText
+  private _mtext?: MTextObject
+  private _text: AcGiMTextData
+  private _style: AcGiTextStyle
 
   constructor(
     text: AcGiMTextData,
@@ -14,21 +17,33 @@ export class AcTrMText extends AcTrEntity {
     styleManager: AcTrStyleManager
   ) {
     super(styleManager)
+    this._text = text
+    this._style = style
+  }
+
+  async draw() {
+    const mtextRenderer = AcTrMTextRenderer.getInstance()
+    if (!mtextRenderer) return
+
     try {
+      const style = this._style
+
       // @ts-expect-error AcGiTextData and MTextData are compatible
-      this._mtext = new MText(text, style, styleManager, FontManager.instance, {
+      this._mtext = await mtextRenderer.renderMText(this._text, style, {
         byLayerColor: style.byLayerColor,
         byBlockColor: style.byBlockColor
       })
-      this.add(this._mtext)
-      this.flatten()
-      this.traverse(object => {
-        // Add the flag to check intersection using bounding box of the mesh
-        object.userData.bboxIntersectionCheck = true
-      })
+      if (this._mtext) {
+        this.add(this._mtext)
+        this.flatten()
+        this.traverse(object => {
+          // Add the flag to check intersection using bounding box of the mesh
+          object.userData.bboxIntersectionCheck = true
+        })
+      }
     } catch (error) {
       console.log(
-        `Failed to render mtext '${text.text}' with the following error:\n`,
+        `Failed to render mtext '${this._text.text}' with the following error:\n`,
         error
       )
     }