+++ packages/three-renderer/src/renderer/index.ts
@@ -1,0 +1,2 @@
+export * from './AcTrMTextRenderer'
+export * from './AcTrRenderer'