+++ packages/three-renderer/src/viewport/AcTrBaseView.ts
@@ -8,7 +8,7 @@ import {
 import * as THREE from 'three'
 import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls'
 
-import { AcTrRenderer } from '../AcTrRenderer'
+import { AcTrRenderer } from '../renderer'
 import { AcTrCamera } from './AcTrCamera'
 
 export interface AcDbEntityEventArgs {